src_py/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anthropic==0.42.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appnope==0.1.3
arfpy==0.1.1
asteroid-filterbanks==0.4.0
asttokens==2.4.1
attrs==23.2.0
autogen-agentchat==0.2.40
beautifulsoup4==4.12.3
boto3==1.34.91
botocore==1.34.91
bs4==0.0.2
build==1.2.2
CacheControl==0.14.0
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
cloudpickle==3.0.0
cobble==0.1.4
colorlog==6.8.2
comm==0.2.1
confluent-kafka==2.3.0
contextlib2==21.6.0
contourpy==1.2.0
copulas==0.11.0
crashtest==0.4.1
ctgan==0.10.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.0
decorator==5.1.1
deepecho==0.6.0
defusedxml==0.7.1
discord.py==2.3.2
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
docker==7.1.0
docopt==0.6.2
dulwich==0.21.7
einops==0.7.0
executing==2.0.1
Faker==24.11.0
fastapi==0.109.0
fastapi-utils==0.2.1
fastjsonschema==2.20.0
filelock==3.13.4
FLAML==2.3.3
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.41
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
HyperPyYAML==1.2.2
idna==3.6
imbalanced-learn==0.12.0
importlib_metadata==8.5.0
installer==0.7.0
ipykernel==6.29.0
ipython==8.21.0
ipywidgets==8.1.5
japanize-matplotlib==1.1.3
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.8.2
jmespath==1.0.1
joblib==1.3.2
jquants-api-client==1.7.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.13
keyring==24.3.1
kiwisolver==1.4.5
langchain==0.3.14
langchain-anthropic==0.3.1
langchain-community==0.3.14
langchain-core==0.3.29
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langgraph==0.2.61
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.48
langsmith==0.2.10
Levenshtein==0.24.0
lightgbm==4.3.0
lightning==2.5.0.post0
lightning-utilities==0.11.9
llvmlite==0.42.0
lxml==5.1.0
Mako==1.3.2
mammoth==1.9.0
Markdown==3.7
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==2.1.4
marshmallow==3.24.1
matplotlib==3.10.0
matplotlib-fontja==1.0.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-collections==0.1.1
more-itertools==10.5.0
mplfinance==0.12.10b0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numba==0.59.0
numpy==1.26.3
oauthlib==3.2.2
omegaconf==2.3.0
open-spiel==1.5
openai==1.59.3
openai-whisper==20240930
optuna==3.5.0
orjson==3.10.13
outcome==1.3.0.post0
packaging==23.2
pandas==1.5.3
parso==0.8.3
pdfminer==20191125
pexpect==4.9.0
pillow==10.2.0
pipenv==2024.4.1
pkginfo==1.11.1
platformdirs==4.2.0
playwright==1.49.1
plotly==5.21.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
polars==1.20.0
primePy==1.3
prompt-toolkit==3.0.43
protobuf==5.28.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyee==12.0.0
Pygments==2.17.2
pyparsing==3.1.1
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytorch-metric-learning==2.8.1
pytorch-tabnet==4.1.0
pytorch-tabular==1.1.1
pytz==2023.4
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.6.1
rdt==1.12.0
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.10.1
scikit-learn==1.4.0
scipy==1.12.0
sdmetrics==0.14.0
sdv==1.12.1
seaborn==0.13.2
selenium==4.17.2
semver==3.0.4
sentencepiece==0.2.0
shap==0.44.1
shellingham==1.5.4
six==1.16.0
slack_sdk==3.26.2
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.5
speechbrain==1.0.2
SQLAlchemy==1.4.51
stack-data==0.6.3
starlette==0.35.1
sympy==1.13.1
tabulate==0.9.0
tavily-python==0.5.0
tenacity==8.2.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.2.0
tiktoken==0.8.0
tomlkit==0.13.2
torch==2.6.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.6.0
torchmetrics==1.5.2
torchvision==0.20.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
trio==0.24.0
trio-websocket==0.11.1
trove-classifiers==2024.9.12
trueskill==0.4.5
tweepy==4.14.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.2.1
uvicorn==0.27.0.post1
virtualenv==20.26.6
wcwidth==0.2.13
webdriver-manager==4.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wsproto==1.2.0
xattr==1.1.0
yarl==1.9.4
zipp==3.20.2
